# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ConnectSessionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.SessionConnectData'
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  dto.CreateSessionRequest:
    properties:
      events:
        example: message,status
        maxLength: 500
        type: string
      name:
        example: default
        maxLength: 50
        minLength: 1
        type: string
      proxy_url:
        example: http://proxy.example.com:8080
        type: string
      webhook_url:
        example: https://webhook.example.com/whatsapp
        type: string
    required:
    - name
    type: object
  dto.CreateSessionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.SessionCreateData'
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  dto.ErrorInfo:
    properties:
      code:
        example: INVALID_REQUEST
        type: string
      details:
        example: Additional error details
        type: string
      message:
        example: Invalid request parameters
        type: string
    type: object
...
  dto.PairPhoneRequest:
    properties:
      phone:
        example: "5511999999999"
        type: string
    required:
    - phone
    type: object
  dto.PairPhoneResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.PairPhoneResponseData'
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  dto.PairPhoneResponseData:
    properties:
      action:
        example: pair
        type: string
      code:
        example: "123456"
        type: string
      phone:
        example: "5511999999999"
        type: string
      session_id:
        example: session_123
        type: string
      status:
        example: success
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  dto.SessionConnectData:
    properties:
      action:
        example: connect
        type: string
      connection:
        $ref: '#/definitions/dto.SessionConnectionInfo'
      qr_code:
        type: string
      session:
        $ref: '#/definitions/dto.SessionInfo'
      session_id:
        example: session_123
        type: string
      status:
        example: success
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  dto.SessionConnectionInfo:
    properties:
      connected:
        type: boolean
      is_connected:
        type: boolean
      qr_code:
        type: string
    type: object
  dto.SessionCreateData:
    properties:
      action:
        example: create
        type: string
      session:
        $ref: '#/definitions/dto.SessionInfo'
      status:
        example: success
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  dto.SessionData:
    properties:
      action:
        example: create
        type: string
      qr_code:
        type: string
      session:
        $ref: '#/definitions/dto.SessionInfo'
      session_id:
        example: session_123
        type: string
      sessions:
        items:
          $ref: '#/definitions/dto.SessionInfo'
        type: array
      status:
        example: success
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  dto.SessionInfo:
    properties:
      api_key:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      device_jid:
        example: 5511999999999@s.whatsapp.net
        type: string
      events:
        example:
        - message
        - status
        items:
          type: string
        type: array
      id:
        example: default
        type: string
      name:
        example: default
        type: string
      proxy_url:
        example: http://proxy.example.com:8080
        type: string
      status:
        example: connected
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      webhook_url:
        example: https://webhook.example.com/whatsapp
        type: string
    type: object
  dto.SessionListResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.SessionData'
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  dto.SessionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.SessionData'
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  dto.SessionStatusResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.SessionStatusResponseData'
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  dto.SessionStatusResponseData:
    properties:
      action:
        example: status
        type: string
      client_status:
        example: connected
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      device_jid:
        example: 5511999999999@s.whatsapp.net
        type: string
      is_connected:
        type: boolean
      name:
        example: My Session
        type: string
      session_id:
        example: session_123
        type: string
      session_status:
        example: connected
        type: string
      status:
        example: success
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  dto.UpdateWebhookRequest:
    properties:
      events:
        example:
        - Message
        - Receipt
        items:
          type: string
        type: array
      secret:
        example: new_webhook_secret_key
        type: string
      status:
        example: active
        type: string
      url:
        example: https://webhook.example.com/meow
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@meow.com
    name: meow API Support
    url: https://github.com/your-username/meow
  description: A meow API server built with Go, inspired by meow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: meow meow API
  version: "1.0"
paths:
...
  /sessions/{id}/connect:
    post:
      consumes:
      - application/json
      description: Initiates connection to meow and returns QR code if needed. Accepts
        session ID or name.
      parameters:
      - description: Session ID or Name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Connection initiated successfully
          schema:
            $ref: '#/definitions/dto.ConnectSessionResponse'
        "400":
          description: Invalid session ID or name
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.SessionResponse'
      security:
      - ApiKeyAuth: []
      summary: Connect a session to meow
      tags:
      - Sessions
  /sessions/{id}/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a meow session and stops the client
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session deleted successfully
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.SessionResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a session
      tags:
      - Sessions
  /sessions/{id}/disconnect:
    post:
      consumes:
      - application/json
      description: Disconnects the session from meow without deleting it
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session disconnected successfully
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.SessionResponse'
      security:
      - ApiKeyAuth: []
      summary: Disconnect a session from meow
      tags:
      - Sessions
  /sessions/{id}/info:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific session by ID or
        name
      parameters:
      - description: Session ID or Name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session information retrieved successfully
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "400":
          description: Invalid session ID or name
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.SessionResponse'
      security:
      - ApiKeyAuth: []
      summary: Get session information
      tags:
      - Sessions
  /sessions/{id}/pair:
    post:
      consumes:
      - application/json
      description: Pairs a phone number with the session for meow connection
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Phone pairing request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PairPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Phone pairing initiated successfully
          schema:
            $ref: '#/definitions/dto.PairPhoneResponse'
        "400":
          description: Invalid request body or session ID
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.SessionResponse'
      security:
      - ApiKeyAuth: []
      summary: Pair phone number with session
      tags:
      - Sessions
  /sessions/{id}/status:
    get:
      consumes:
      - application/json
      description: Retrieves the current status and connection information of a session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session status retrieved successfully
          schema:
            $ref: '#/definitions/dto.SessionStatusResponse'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.SessionResponse'
      security:
      - ApiKeyAuth: []
      summary: Get session status
      tags:
      - Sessions
  /sessions/{id}/webhook:
    put:
      consumes:
      - application/json
      description: Updates the webhook URL and events for a session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook updated successfully
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.SessionResponse'
      security:
      - ApiKeyAuth: []
      summary: Update session webhook URL
      tags:
      - Sessions
  /sessions/create:
    post:
      consumes:
      - application/json
      description: Creates a new meow session and starts the client
      parameters:
      - description: Session creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Session created successfully
          schema:
            $ref: '#/definitions/dto.CreateSessionResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.SessionResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new meow session
      tags:
      - Sessions
  /sessions/list:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all meow sessions
      produces:
      - application/json
      responses:
        "200":
          description: Sessions retrieved successfully
          schema:
            $ref: '#/definitions/dto.SessionListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.SessionResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all sessions
      tags:
      - Sessions
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key authentication. Simply provide your API key directly: "YOUR_API_KEY".
      The system automatically detects if it''s a Global API Key (can access all sessions
      and session management) or a Session-specific API Key (can only access the specific
      session it belongs to).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...